+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Undeployed
 
 * Fix capitalization bug for new bulletins
+* Bug fix: link type bulletins do not save body attribute
 
 Version 2.1.6 - May 23, 2013